# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wending Planner API",
        "description": "API de gestion pour la préparation du mariage.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/seating-tables": {
            "get": {
                "tags": [
                    "SeatingTables"
                ],
                "summary": "Liste des tables du plan de table, avec leurs invités",
                "operationId": "0bdffd1a23c686921d6dc843766a9a1b",
                "responses": {
                    "200": {
                        "description": "Liste des tables, par ordre alphabétique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SeatingTable"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SeatingTables"
                ],
                "summary": "Créer une table",
                "operationId": "e526ffcefdde4695920e4e121e162689",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeatingTableInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Table créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeatingTable"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/seating-tables/{seating_table}": {
            "get": {
                "tags": [
                    "SeatingTables"
                ],
                "summary": "Détail d'une table",
                "operationId": "e4f0b501b2f6d50fca3a8715fbd26843",
                "parameters": [
                    {
                        "name": "seating_table",
                        "in": "path",
                        "description": "ID de la table",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeatingTable"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "SeatingTables"
                ],
                "summary": "Mettre à jour une table",
                "operationId": "40c8c9f788ff3ac802d34de61ac0ce76",
                "parameters": [
                    {
                        "name": "seating_table",
                        "in": "path",
                        "description": "ID de la table",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeatingTableInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeatingTable"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SeatingTables"
                ],
                "summary": "Supprimer une table (les invités redeviennent non assignés)",
                "operationId": "ecb60d2afb4d4a55a66e9209a8c717cf",
                "parameters": [
                    {
                        "name": "seating_table",
                        "in": "path",
                        "description": "ID de la table",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeatingTable"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Table introuvable"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Liste des tâches",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Liste des tâches, triées par échéance (sans échéance en dernier)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Créer une tâche",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tâche créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Détail d'une tâche",
                "operationId": "9736e8dac4ca267e196496c4b790c757",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID de la tâche",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tâche trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mettre à jour une tâche",
                "operationId": "4732012736ffb7728b7e6b7949d3c140",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID de la tâche",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tâche mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Supprimer une tâche",
                "operationId": "3aa12cf460ff0476704f02693959d7de",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID de la tâche",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tâche supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tâche introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SeatingTable": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Table 1"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 8
                    },
                    "guests": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "confirmed": {
                                    "type": "boolean",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SeatingTableInput": {
                "required": [
                    "name",
                    "capacity"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Table 1",
                        "maxLength": 255
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 8,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "Task": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Réserver la salle"
                    },
                    "category": {
                        "type": "string",
                        "example": "prestataires",
                        "nullable": true,
                        "enum": [
                            "administratif",
                            "prestataires",
                            "tenues",
                            "deco",
                            "invitations",
                            "beaute",
                            "logistique",
                            "autre"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "todo",
                        "enum": [
                            "todo",
                            "in_progress",
                            "done"
                        ]
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2027-03-15",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TaskInput": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Réserver la salle",
                        "maxLength": 255
                    },
                    "category": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "administratif",
                            "prestataires",
                            "tenues",
                            "deco",
                            "invitations",
                            "beaute",
                            "logistique",
                            "autre"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "todo",
                            "in_progress",
                            "done"
                        ]
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}